day09 点赞功能

+++ heima-leadnews-model/src/main/java/com/heima/model/behavior/dtos/LikesBehaviorDto.java
@@ -14,7 +14,7 @@ import lombok.Data;
 public class LikesBehaviorDto {
     //文章id
     @IdEncrypt
-    private Integer articleId;
+    private Long articleId;
     //操作类型 0 点赞 1 取消点赞
     private Short operation;
     //类型 0文章 1动态 2评论
